Add files via upload
We categorized all studies into two groups: cases with non-constant flow on the main stream only, and cases with simultaneous non-constant flow on both the main and tributary streams. The initial water level (h0) in the channel is set to 0.01 m, and the confluence angle is 60°. For cases with non-constant flow on the main stream only, the constant flow on the tributary is set to 22.5×10-3 m³/s, with variables including the non-constant flow period (T) and the non-constant flow variability (F). In cases with simultaneous non-constant flow on both the main stream and tributary, the variables under investigation are the staggered period of the two non-constant streams and the non-constant flow variability (F) on the main stream.